Move FacultyCarousel class to its own file
